create new validator flag to check whether it is registered already and extract the staking contract related tests in a separate file;
