+++ contracts/Hydra/common/Errors.sol
@@ -4,3 +4,4 @@ pragma solidity 0.8.17;
 error Unauthorized(string only);
 error StakeRequirement(string src, string msg);
 error InvalidSignature(address signer);
+error AlreadyRegistered(address validator);
