+++ hardhat.config.ts
@@ -9,7 +9,7 @@ import "./big-int-fix.ts";
 dotenv.config();
 
 // eslint-disable-next-line import/first
-import "./tasks";
+// import "./tasks";
 
 const config: HardhatUserConfig = {
   solidity: {
@@ -71,7 +71,7 @@ const config: HardhatUserConfig = {
   },
   dodoc: {
     // uncomment to stop docs from autogenerating each compile
-    // runOnCompile: false,
+    runOnCompile: false,
   },
 };
